pyproject.toml
[tool.poetry]
name = "kong-batch"
version = "0.9.3"
description = "Batch job submission and management tool"
authors = ["Paul Gessinger <hello@paulgessinger.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/paulgessinger/kong/"
packages = [
    { include = "kong", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.1.2"
pyyaml = "^5.3.1"
halo = ">=0.0.29,<0.0.31"
sh = "^1.13.1"
python-dateutil = "^2.8.1"
peewee = "^3.13.3"
coloredlogs = "^14.0"
humanfriendly = "^8.2"
psutil = "^5.7.2"
tqdm = "^4.48.0"
notifiers = "^1.2.1"
jinja2 = "^2.11.2"
pytimeparse = "^1.1.8"
ipython = {version = "^7.16.1", optional = true}
pytz = "^2020.1"
pydantic = "^1.6.1"

[tool.poetry.scripts]
kong = 'kong.cli:main'

[tool.poetry.extras]
ipython = ["ipython"]

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
coverage = "^5.2"
pytest-cov = "^2.10.1"
pytest-rerunfailures = "^9.0"
mypy = "^0.782"
flake8 = "^3.8.3"
black = "^19.10b0"
tox = "^3.19.0"
sphinx-autodoc-typehints = "^1.11.0"

[tool.semantic_release]
version_pattern = [
    'pyproject.toml:version = "{version}"',
]
commit_subject = "Bump and tag {version}"
commit_message = ""
build_command = "pip install poetry && poetry build"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
